+++ minecraft/src/net/minecraft/client/resources/sounds/AbstractSoundInstance.java
@@ -15,9 +15,9 @@ public abstract class AbstractSoundInstance implements SoundInstance {
 	protected final ResourceLocation location;
 	protected float volume = 1.0F;
 	protected float pitch = 1.0F;
-	protected float x;
-	protected float y;
-	protected float z;
+	protected double x;
+	protected double y;
+	protected double z;
 	protected boolean looping;
 	protected int delay;
 	protected SoundInstance.Attenuation attenuation = SoundInstance.Attenuation.LINEAR;
@@ -81,17 +81,17 @@ public abstract class AbstractSoundInstance implements SoundInstance {
 	}
 
 	@Override
-	public float getX() {
+	public double getX() {
 		return this.x;
 	}
 
 	@Override
-	public float getY() {
+	public double getY() {
 		return this.y;
 	}
 
 	@Override
-	public float getZ() {
+	public double getZ() {
 		return this.z;
 	}
