+++ minecraft/src/net/minecraft/client/resources/sounds/MinecartSoundInstance.java
@@ -19,9 +19,9 @@ public class MinecartSoundInstance extends AbstractTickableSoundInstance {
 		this.looping = true;
 		this.delay = 0;
 		this.volume = 0.0F;
-		this.x = (float)abstractMinecart.getX();
-		this.y = (float)abstractMinecart.getY();
-		this.z = (float)abstractMinecart.getZ();
+		this.x = (double)((float)abstractMinecart.getX());
+		this.y = (double)((float)abstractMinecart.getY());
+		this.z = (double)((float)abstractMinecart.getZ());
 	}
 
 	@Override
@@ -39,9 +39,9 @@ public class MinecartSoundInstance extends AbstractTickableSoundInstance {
 		if (this.minecart.removed) {
 			this.stop();
 		} else {
-			this.x = (float)this.minecart.getX();
-			this.y = (float)this.minecart.getY();
-			this.z = (float)this.minecart.getZ();
+			this.x = (double)((float)this.minecart.getX());
+			this.y = (double)((float)this.minecart.getY());
+			this.z = (double)((float)this.minecart.getZ());
 			float f = Mth.sqrt(Entity.getHorizontalDistanceSqr(this.minecart.getDeltaMovement()));
 			if ((double)f >= 0.01) {
 				this.pitch = Mth.clamp(this.pitch + 0.0025F, 0.0F, 1.0F);
