+++ minecraft/src/net/minecraft/world/entity/ai/navigation/PathNavigation.java
@@ -210,9 +210,9 @@ public abstract class PathNavigation {
 		Vec3 vec3 = this.getTempMobPos();
 		this.maxDistanceToWaypoint = this.mob.getBbWidth() > 0.75F ? this.mob.getBbWidth() / 2.0F : 0.75F - this.mob.getBbWidth() / 2.0F;
 		Vec3i vec3i = this.path.currentPos();
-		double d = Math.abs(this.mob.getX() - (double)((float)vec3i.getX() + 0.5F));
+		double d = Math.abs(this.mob.getX() - ((double)vec3i.getX() + 0.5));
 		double e = Math.abs(this.mob.getY() - (double)vec3i.getY());
-		double f = Math.abs(this.mob.getZ() - (double)((float)vec3i.getZ() + 0.5F));
+		double f = Math.abs(this.mob.getZ() - ((double)vec3i.getZ() + 0.5));
 		boolean bl = d < (double)this.maxDistanceToWaypoint && f < (double)this.maxDistanceToWaypoint && e < 1.0;
 		if (bl || this.mob.canCutCorner(this.path.currentNode().type) && this.shouldTargetNextNodeInDirection(vec3)) {
 			this.path.setIndex(this.path.getIndex() + 1);
