+++ minecraft/src/net/minecraft/world/level/block/CampfireBlock.java
@@ -162,9 +162,9 @@ public class CampfireBlock extends BaseEntityBlock implements SimpleWaterloggedB
 		if ((Boolean)blockState.getValue(LIT)) {
 			if (random.nextInt(10) == 0) {
 				level.playLocalSound(
-					(double)((float)blockPos.getX() + 0.5F),
-					(double)((float)blockPos.getY() + 0.5F),
-					(double)((float)blockPos.getZ() + 0.5F),
+					(double)blockPos.getX() + 0.5,
+					(double)blockPos.getY() + 0.5,
+					(double)blockPos.getZ() + 0.5,
 					SoundEvents.CAMPFIRE_CRACKLE,
 					SoundSource.BLOCKS,
 					0.5F + random.nextFloat(),
@@ -177,9 +177,9 @@ public class CampfireBlock extends BaseEntityBlock implements SimpleWaterloggedB
 				for (int i = 0; i < random.nextInt(1) + 1; i++) {
 					level.addParticle(
 						ParticleTypes.LAVA,
-						(double)((float)blockPos.getX() + 0.5F),
-						(double)((float)blockPos.getY() + 0.5F),
-						(double)((float)blockPos.getZ() + 0.5F),
+						(double)blockPos.getX() + 0.5,
+						(double)blockPos.getY() + 0.5,
+						(double)blockPos.getZ() + 0.5,
 						(double)(random.nextFloat() / 2.0F),
 						5.0E-5,
 						(double)(random.nextFloat() / 2.0F)
