+++ minecraft/src/net/minecraft/world/level/block/EndGatewayBlock.java
@@ -32,12 +32,12 @@ public class EndGatewayBlock extends BaseEntityBlock {
 			int i = ((TheEndGatewayBlockEntity)blockEntity).getParticleAmount();
 
 			for (int j = 0; j < i; j++) {
-				double d = (double)blockPos.getX() + (double)random.nextFloat();
-				double e = (double)blockPos.getY() + (double)random.nextFloat();
-				double f = (double)blockPos.getZ() + (double)random.nextFloat();
-				double g = ((double)random.nextFloat() - 0.5) * 0.5;
-				double h = ((double)random.nextFloat() - 0.5) * 0.5;
-				double k = ((double)random.nextFloat() - 0.5) * 0.5;
+				double d = (double)blockPos.getX() + random.nextDouble();
+				double e = (double)blockPos.getY() + random.nextDouble();
+				double f = (double)blockPos.getZ() + random.nextDouble();
+				double g = (random.nextDouble() - 0.5) * 0.5;
+				double h = (random.nextDouble() - 0.5) * 0.5;
+				double k = (random.nextDouble() - 0.5) * 0.5;
 				int l = random.nextInt(2) * 2 - 1;
 				if (random.nextBoolean()) {
 					f = (double)blockPos.getZ() + 0.5 + 0.25 * (double)l;
