+++ minecraft/src/net/minecraft/world/level/block/FallingBlock.java
@@ -70,9 +70,9 @@ public class FallingBlock extends Block {
 		if (random.nextInt(16) == 0) {
 			BlockPos blockPos2 = blockPos.below();
 			if (isFree(level.getBlockState(blockPos2))) {
-				double d = (double)blockPos.getX() + (double)random.nextFloat();
+				double d = (double)blockPos.getX() + random.nextDouble();
 				double e = (double)blockPos.getY() - 0.05;
-				double f = (double)blockPos.getZ() + (double)random.nextFloat();
+				double f = (double)blockPos.getZ() + random.nextDouble();
 				level.addParticle(new BlockParticleOption(ParticleTypes.FALLING_DUST, blockState), d, e, f, 0.0, 0.0, 0.0);
 			}
 		}
