+++ minecraft/src/net/minecraft/world/level/block/LeavesBlock.java
@@ -101,9 +101,9 @@ public class LeavesBlock extends Block {
 				BlockPos blockPos2 = blockPos.below();
 				BlockState blockState2 = level.getBlockState(blockPos2);
 				if (!blockState2.canOcclude() || !blockState2.isFaceSturdy(level, blockPos2, Direction.UP)) {
-					double d = (double)((float)blockPos.getX() + random.nextFloat());
+					double d = (double)blockPos.getX() + random.nextDouble();
 					double e = (double)blockPos.getY() - 0.05;
-					double f = (double)((float)blockPos.getZ() + random.nextFloat());
+					double f = (double)blockPos.getZ() + random.nextDouble();
 					level.addParticle(ParticleTypes.DRIPPING_WATER, d, e, f, 0.0, 0.0, 0.0);
 				}
 			}
