+++ minecraft/src/net/minecraft/world/level/block/entity/EnchantmentTableBlockEntity.java
@@ -51,13 +51,7 @@ public class EnchantmentTableBlockEntity extends BlockEntity implements Nameable
 		this.oOpen = this.open;
 		this.oRot = this.rot;
 		Player player = this.level
-			.getNearestPlayer(
-				(double)((float)this.worldPosition.getX() + 0.5F),
-				(double)((float)this.worldPosition.getY() + 0.5F),
-				(double)((float)this.worldPosition.getZ() + 0.5F),
-				3.0,
-				false
-			);
+			.getNearestPlayer((double)this.worldPosition.getX() + 0.5, (double)this.worldPosition.getY() + 0.5, (double)this.worldPosition.getZ() + 0.5, 3.0, false);
 		if (player != null) {
 			double d = player.getX() - ((double)this.worldPosition.getX() + 0.5);
 			double e = player.getZ() - ((double)this.worldPosition.getZ() + 0.5);
